+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  * Mixed rust/python layout
+ * Added PEP 517 support
 
 ## [0.6.1]